+++ src/class.mjs
@@ -1,0 +1,1 @@
+export { default as get_class } from './class/get_class.mjs';